requirements.txt
docx2txt>=0.8
langchain-core>=0.1.0
langchain-openai>=0.0.5
python-dotenv>=1.0.0
python-docx>=0.8.11
pandas>=2.0.0
openpyxl>=3.1.2